pyproject.toml
[tool.poetry]
name = "flask-boilerplate"
version = "0.1.0"
description = "Flask BoilerPlate"
authors = ["Zeeshan Asim <zeeshanravian1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^4.0.1"
types-flask-cors = "^4.0.0.20240405"
environs = "^11.0.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pre-commit = "^3.7.0"
python-lsp-server = {extras = ["all"], version = "^1.11.0"}
black = "^24.4.2"
ipykernel = "^6.29.4"
jupyter = "^1.0.0"
radon = "^6.0.1"
ruff = "^0.4.3"
skjold = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.skjold]
sources = ["pyup"]                      # Sources to check against.
report_only = false                     # Report only, always exit with zero.
report_format = 'cli'                   # Output findings as `json`. Default is 'cli'.
cache_dir = '.skjold_cache'             # Cache location (default: `~/.skjold/cache`).
cache_expires = 86400                   # Cache max. age.
ignore_file = '.skjoldignore'           # Ignorefile location (default `.skjoldignore`).
verbose = true
